# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "JobBit APIs",
    "version": "v1"
  },
  "servers": [
    { }
  ],
  "paths": {
...
    "/api/Requests/GetRequestByID/{RequestID}": {
      "get": {
        "tags": [
          "Requests"
        ],
        "operationId": "GetRequestByID",
        "parameters": [
          {
            "name": "RequestID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AllRequestInfo"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllRequestInfo"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllRequestInfo"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AllCompanyInfo": {
        "type": "object",
        "properties": {
          "companyID": {
            "type": "integer",
            "format": "int32"
          },
          "wilayaInfo": {
            "$ref": "#/components/schemas/WilayaDTO"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "logoPath": {
            "type": "string",
            "nullable": true
          },
          "link": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "AllJobInfoDTO": {
        "type": "object",
        "properties": {
          "jobID": {
            "type": "integer",
            "format": "int32"
          },
          "comapnyInfo": {
            "$ref": "#/components/schemas/AllCompanyInfo"
          },
          "jobType": {
            "type": "string",
            "nullable": true
          },
          "postedDate": {
            "type": "string",
            "format": "date-time"
          },
          "experience": {
            "type": "string",
            "nullable": true
          },
          "available": {
            "type": "boolean"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "skills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkillDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AllJobSeekerInfo": {
        "type": "object",
        "properties": {
          "jobSeekerID": {
            "type": "integer",
            "format": "int32"
          },
          "wilayaInfo": {
            "$ref": "#/components/schemas/WilayaDTO"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "profilePicturePath": {
            "type": "string",
            "nullable": true
          },
          "cvFilePath": {
            "type": "string",
            "nullable": true
          },
          "linkProfileLinkden": {
            "type": "string",
            "nullable": true
          },
          "linkProfileGithub": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "skills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkillDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AllRequestInfo": {
        "type": "object",
        "properties": {
          "requestID": {
            "type": "integer",
            "format": "int32"
          },
          "jobSeekerInfo": {
            "$ref": "#/components/schemas/AllJobSeekerInfo"
          },
          "jobInfo": {
            "$ref": "#/components/schemas/AllJobInfoDTO"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "boolean",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "SkillDTO": {
        "type": "object",
        "properties": {
          "skillID": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "iconUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "WilayaDTO": {
        "type": "object",
        "properties": {
          "wilayaID": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}